feat: add support for directory based rendering of JSX
