+++ apps/juxtaposition-ui/src/services/juxt-web/routes/console/messages.jsx
@@ -1,5 +1,5 @@
 import { buildContext } from '@/services/juxt-web/views/context';
-import { MessagesView } from '@/services/juxt-web/views/messages';
+import { WebMessagesView } from '@/services/juxt-web/views/web/messages';
 
 const crypto = require('crypto');
 const express = require('express');
@@ -14,8 +14,10 @@ const router = express.Router();
 
 router.get('/', async function (req, res) {
 	const conversations = await database.getConversations(req.pid);
-	res.jsx(<MessagesView conversations={conversations} ctx={buildContext(res)}  />);
-});	
+	res.jsxForDirectory({
+		web: <WebMessagesView conversations={conversations} ctx={buildContext(res)} />,
+	});
+});
 
 router.post('/new', async function (req, res) {
 	let conversation = await database.getConversationByID(req.body.community_id);
